sigs-django/requirements.txt
Django==1.10.5
django_mysql==2.1.0
mysqlclient==1.3.7
redis==2.10.5
pyjwt==1.4.2
coreapi
numpy
netCDF4
pandas
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
httpie
django-filter
django-cors-middleware==1.3.1
